Core/Entities: rename ItemPrototype to ItemTemplate
