+++ src/server/game/Entities/Item/ItemTemplate.cpp
@@ -15,7 +15,7 @@
 * with this program. If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "ItemPrototype.h"
+#include "ItemTemplate.h"
 
 bool ItemTemplate::CanChangeEquipStateInCombat() const
 {
