+++ src/server/game/Entities/Item/ItemTemplate.h
